Moved user management to free tier
