+++ backend/src/services/user/userEditor.ts
@@ -1,11 +1,11 @@
 import assert from 'assert'
-import Roles from '../../../security/roles'
-import Error400 from '../../../errors/Error400'
-import SequelizeRepository from '../../../database/repositories/sequelizeRepository'
-import UserRepository from '../../../database/repositories/userRepository'
-import TenantUserRepository from '../../../database/repositories/tenantUserRepository'
-import Plans from '../../../security/plans'
-import { IServiceOptions } from '../../IServiceOptions'
+import Roles from '../../security/roles'
+import Error400 from '../../errors/Error400'
+import SequelizeRepository from '../../database/repositories/sequelizeRepository'
+import UserRepository from '../../database/repositories/userRepository'
+import TenantUserRepository from '../../database/repositories/tenantUserRepository'
+import Plans from '../../security/plans'
+import { IServiceOptions } from '../IServiceOptions'
 
 /**
  * Handles the edition of the user(s) via the User page.
