+++ frontend/src/middleware/auth/permission-guard.js
@@ -1,4 +1,4 @@
-import { PermissionChecker } from '@/premium/user/permission-checker'
+import { PermissionChecker } from '@/modules/user/permission-checker'
 
 /**
  * Permission Guard
