+++ frontend/src/modules/auth/pages/password-reset-page.vue
@@ -155,7 +155,7 @@
 
 <script>
 import { mapGetters, mapActions } from 'vuex'
-import { UserModel } from '@/premium/user/user-model'
+import { UserModel } from '@/modules/user/user-model'
 import AppI18n from '@/shared/i18n/i18n'
 import { passwordConfirmRules } from '@/modules/auth/auth-helpers'
 const { fields } = UserModel
