+++ frontend/src/modules/conversation/conversation-permissions.js
@@ -1,5 +1,5 @@
 import Permissions from '@/security/permissions'
-import { PermissionChecker } from '@/premium/user/permission-checker'
+import { PermissionChecker } from '@/modules/user/permission-checker'
 
 export class ConversationPermissions {
   constructor(currentTenant, currentUser) {
