+++ frontend/src/modules/index.js
@@ -17,7 +17,7 @@ import task from '@/modules/task/task-module'
 
 import communityHelpCenter from '@/premium/community-help-center/community-help-center-module'
 import eagleEye from '@/premium/eagle-eye/eagle-eye-module'
-import user from '@/premium/user/user-module'
+import user from '@/modules/user/user-module'
 
 const modules = {
   shared,
