+++ frontend/src/modules/task/task-permissions.js
@@ -1,5 +1,5 @@
 import Permissions from '@/security/permissions'
-import { PermissionChecker } from '@/premium/user/permission-checker'
+import { PermissionChecker } from '@/modules/user/permission-checker'
 
 export class TaskPermissions {
   constructor(currentTenant, currentUser) {
