+++ frontend/src/modules/user/components/form/user-form.vue
@@ -116,7 +116,7 @@
 
 <script>
 import { FormSchema } from '@/shared/form/form-schema'
-import { UserModel } from '@/premium/user/user-model'
+import { UserModel } from '@/modules/user/user-model'
 import { i18n } from '@/i18n'
 import Message from '@/shared/message/message'
 import config from '@/config'
