+++ frontend/src/modules/user/permission-checker.js
