+++ frontend/src/modules/user/user-form-store.js
@@ -1,7 +1,7 @@
 import Errors from '@/shared/error/errors'
 import Message from '@/shared/message/message'
 import { i18n } from '@/i18n'
-import { UserService } from '@/premium/user/user-service'
+import { UserService } from '@/modules/user/user-service'
 
 export default {
   namespaced: true,
