+++ frontend/src/modules/user/user-list-exporter-fields.js
@@ -1,4 +1,4 @@
-import { UserModel } from '@/premium/user/user-model'
+import { UserModel } from '@/modules/user/user-model'
 
 const { fields } = UserModel
