+++ frontend/src/modules/user/user-module.js
@@ -1,0 +1,7 @@
+import routes from '@/modules/user/user-routes'
+import store from '@/modules/user/user-store'
+
+export default {
+  routes,
+  store
+}
