--- frontend/src/premium/user/user-module.js
@@ -1,7 +1,0 @@
-import routes from '@/premium/user/user-routes'
-import store from '@/premium/user/user-store'
-
-export default {
-  routes,
-  store
-}
