+++ frontend/src/shared/list/list-item-relation-to-one.vue
@@ -11,7 +11,7 @@
 
 <script>
 import { mapGetters } from 'vuex'
-import { PermissionChecker } from '@/premium/user/permission-checker'
+import { PermissionChecker } from '@/modules/user/permission-checker'
 
 export default {
   name: 'AppListItemRelationToOne',
